add remove

+++ imagitor/src/hook/useItem.ts
@@ -106,17 +106,17 @@ const useItem = () => {
     }
   };
 
-  const removeItem = (targetItemId: string | string[]) => {
+  const removeItem = (targetItemId: string | string[], sendToServer = true) => {
     dispatch(stageDataAction.removeItem(targetItemId));
     console.log(targetItemId);
-    if (page.current !== -1) {
+    if (sendToServer && page.current !== -1) {
       fetch(`${SERVER_URL}/api/projects/${page.current}/item`, {
         method: "DELETE",
         headers: {
           "Authorization": `Bearer ${auth.me.accessToken}`,
           "Content-Type": "application/json"
         },
-        body: JSON.stringify({ updatedObjectId: targetItemId[0] }),
+        body: JSON.stringify({ updatedObjectId: targetItemId }),
       });
     }
   };
